requirements.txt
fastapi
uvicorn
langchain
pypdf
pinecone-client
sentence-transformers
langchain-pinecone
langchain-groq
python-dotenv